Revert "Fix Python ASGI adaptor to handle streaming responses correctly"
