+++ src/pyodide/internal/serializeJsModule.ts
@@ -67,7 +67,7 @@ export function createImportProxy(
   if (!IS_CREATING_SNAPSHOT) {
     return mod;
   }
-  if (!mod || (typeof mod !== 'object' && typeof mod !== 'function')) {
+  if (!mod || typeof mod !== 'object') {
     return mod;
   }
   return new Proxy(mod, {
