+++ src/workerd/server/tests/python/BUILD.bazel
@@ -25,6 +25,8 @@ py_wd_test("env-param")
 
 py_wd_test("asgi")
 
+py_wd_test("asgi-sse")
+
 py_wd_test("random")
 
 py_wd_test("subdirectory")
