+++ src/workerd/server/tests/python/asgi-sse/asgi-sse.wd-test
@@ -1,0 +1,17 @@
+using Workerd = import "/workerd/workerd.capnp";
+
+const unitTests :Workerd.Config = (
+  services = [
+    ( name = "python-asgi-sse",
+      worker = (
+        modules = [
+          (name = "worker.py", pythonModule = embed "worker.py"),
+        ],
+        bindings = [
+          ( name = "SELF", service = "python-asgi-sse" ),
+        ],
+        compatibilityFlags = [%PYTHON_FEATURE_FLAGS, "python_no_global_handlers"],
+      )
+    )
+  ],
+);
